pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-prix"
version = "0.2.0"
description = "Multi-LLM comparison tool for LM Studio"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "gradio>=4.0.0",
    "httpx>=0.25.0",
    "local-inference-pool>=0.1.0",
    "huggingface_hub>=0.25.0",
    "lmstudio>=1.0.0",
    "mcp>=1.20.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "respx>=0.20.0",
]
[project.scripts]
prompt-prix = "prompt_prix.main:run"
prompt-prix-mcp = "prompt_prix.mcp.server:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["prompt_prix*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v -m 'not integration'"
markers = [
    "integration: marks tests that require external services (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["prompt_prix"]
omit = ["tests/*"]
